// api_services/business/business.interface.ts
export interface SingleShopType {
  id: number;
  title: string;
  address: string;
  activity: Activity;
  rate: string;
  rate_count: number;
  location: Location;
  post_code: string;
  email: string;
  logo_location: string;
  has_deposit: boolean;
  send_price: number;
  send_type: string;
  coverage_post_codes: null[];
  commission_percentage: string;
  username: string;
  distance: number;
  created_at: string;
  categories: Category[];
  tags: Category[];
  info: Info;
}

export interface Category {
  id: number;
  title: string;
  image_location: string;
  created_at: string;
  type: "primary" | "secondary";
}

export interface Info {
  id: number;
  start_contract_date: string;
  start_express_contract_date: string;
  express_contract: boolean;
  express_contract_commission: number;
  ceo_first_name: string;
  ceo_last_name: string;
  ceo_mobile_number: string;
  ceo_address: string;
  ceo_email: string;
  created_at: null;
}

export interface Location {
  latitude: string;
  longitude: string;
}

export interface FreeLancerDto {
  url?: string;
  body?: { [key: string]: string | number };
}

// Generated by https://quicktype.io
export interface StroesDto {
  data: StroesDataDto[];
  meta: MetaType;
}

export interface StroesDataDto {
  id: number;
  activity: Activity;
  title: string;
  address: string;
  rate: string;
  location: Location;
  post_code: string;
  email: string;
  logo_location: string;
  has_deposit: boolean;
  send_price: number;
  send_type: string;
  coverage_post_codes: string[];
  commission_percentage: string;
  username: string;
  distance: number;
  created_at: string;
  categories: Category[];
  info: Info;
}
// Generated by https://quicktype.io

export interface Activity {
  from: string;
  to: string;
  day_name: string;
}
// ...
export interface ProductsCarddto {
  data: ProductsCardDatadto[];
  meta: MetaType;
}

export interface MetaType {
  current_page: number;
  from: number;
  last_page: number;
  links: Link[];
  path: string;
  per_page: number;
  to: number;
  total: number;
}

export interface Link {
  url: null | string;
  label: string;
  active: boolean;
}

export interface ProductsCardDatadto {
  id: number;
  in_cart?: number;
  is_favorite?: boolean;
  unit_value: number;
  buy_price: number;
  sell_price: number;
  discount_amount: number;
  discount_percentage: string;
  view_count: number;
  unit: Unit;
  product: Product;
  business: Business;
}

export interface Business {
  id: number;
  title: string;
  address: string;
  rate: string;
  location: Location;
  post_code: string;
  email: string;
  logo_location: string;
  has_deposit: boolean;
  send_price: number;
  send_type: string;
  coverage_post_codes: string[];
  commission_percentage: string;
  username: string;
  created_at: string;
}

export interface Location {
  latitude: string;
  longitude: string;
}

export interface Product {
  id: number;
  title: string;
  description: string;
  cover_location: CoverLocation;
  unit_value: number;
  code: string;
  barcode: string;
  rate: string;
  buy_price: number;
  sell_price: number;
  discount_amount: number;
  discount_percentage: string;
  media_files: CoverLocation[];
  depositItem: DepositItem;
}

export interface CoverLocation {
  id: string;
  sort_number: null;
  file_location: string;
  file_mime_type: string;
}

export interface DepositItem {
  id: number;
  title: string;
  deposit_price?: number;
  image_location?: string | null;
}

export interface Unit {
  id: number;
  title: string;
  key: string;
}

export type SendBusinessCommentDto = {
  rate: string | null | number;
  title: string | null;
  message: string | null;
};

// Generated by https://quicktype.io

export interface SingleBusinessComments {
  data: Datum[];
  links: Links;
  meta: Meta;
  status: string;
  messages: null;
}

export interface Datum {
  id: number;
  rate: number;
  title: null;
  message: string;
  reply_message: string;
  created_at: string;
  admin: Admin;
  user: User;
  status: Status;
}

export interface Admin {
  name: string;
  family: string;
}

export interface Status {
  id: number;
  step: number;
  title: string;
  description: null;
  color: string;
}

export interface User {
  id: number;
  first_name: null;
  last_name: null;
  mobile_number: string;
  email: null;
  birth_date: null;
  gender: null;
  credit: number;
  zip_codes: null;
  created_at: string;
  updated_at: string;
  email_verified_at: null;
  mobile_verified_at: string;
}

export interface Links {
  first: string;
  last: string;
  prev: null;
  next: null;
}

export interface Meta {
  current_page: number;
  from: number;
  last_page: number;
  links: Link[];
  path: string;
  per_page: number;
  to: number;
  total: number;
}

export interface Link {
  url: null | string;
  label: string;
  active: boolean;
}
// Generated by https://quicktype.io

export interface SingleShopBanners {
  "section-1"?: Section1[];
  "section-2"?: Section1[];
  "section-3"?: Section1[];
}

export interface Section1 {
  id: number;
  category_id: null;
  image_location: string;
  title: string;
  group: string;
  description: string;
  position: string;
  link: string;
  expire_date: string;
  created_at: string;
}
export type bussinessDto = {
  post_code: string | number;
  page?: string | number;
  filter_category?: string;
  q?: string | null;
  category_id?: [string] | string[] | null | string;
};