// api_services/product/product.interface.ts
export interface RateRangeproducts {
  total_count: number;
  rates: Rate[];
}

export interface Rate {
  rate: number;
  count: number;
}